requirements.txt
# Core dependencies
cryptography>=41.0.0
PyYAML>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0

# Python version specific
tomli>=2.0.0; python_version < "3.11"

# AI Provider SDKs (optional but recommended)
openai>=1.3.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Development tools
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Logging and monitoring
structlog>=23.2.0
python-json-logger>=2.0.7

# Type hints
types-PyYAML>=6.0.0
types-requests>=2.31.0
